requirements.txt
Flask==1.1.2
pytest==5.4.3
pylint==2.5.3
flask-socketio==4.3.1
gevent-websocket==0.10.1
python-google-places==1.4.1
google-search-results==1.8.3